backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-cors==0.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
idna==3.7
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
